pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentflow-mcp"
version = "0.1.0-beta.1"
description = "APC + AgentFlow MCP Servers - 100% cost savings with FREE models for Claude Code"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "AgentFlow Contributors", email = "gontrand@local"}
]
maintainers = [
    {name = "AgentFlow Contributors", email = "gontrand@local"}
]
keywords = [
    "mcp",
    "claude",
    "llm",
    "ai",
    "cost-optimization",
    "context-compression",
    "quality-validation",
    "free-models",
    "openrouter"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "anthropic>=0.18.0",
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "structlog>=23.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/GontrandL/agentflow-mcp"
Documentation = "https://github.com/GontrandL/agentflow-mcp#readme"
Repository = "https://github.com/GontrandL/agentflow-mcp"
Changelog = "https://github.com/GontrandL/agentflow-mcp/blob/main/CHANGELOG.md"
Issues = "https://github.com/GontrandL/agentflow-mcp/issues"

[project.scripts]
apc-mcp = "agentflow.mcp.apc_mcp_server:main"
agentflow-mcp = "agentflow.mcp.agentflow_mcp_server:main"

[tool.setuptools]
packages = ["agentflow"]

[tool.setuptools.package-data]
agentflow = ["py.typed"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
pythonpath = ["."]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
